refactor utils/ module into multiple submodules
